+++ app/src/main/java/com/example/proyecto1/ui/theme/screens/ImagesUi.kt
@@ -19,6 +19,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.rememberImagePainter
+
 @Composable
 fun NetworkImage(isHighQuality: Boolean) {
     val imageUrl = if (isHighQuality) {
