--- app/src/main/java/com/example/proyecto1/ui/theme/screens/NetworkMonitorActivity.kt
@@ -1,70 +1,0 @@
-package com.example.proyecto1.ui.theme.screens
-
-import android.Manifest
-import android.content.pm.PackageManager
-import android.net.ConnectivityManager
-import android.net.wifi.WifiManager
-import android.os.Bundle
-import android.widget.Toast
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.result.contract.ActivityResultContracts.RequestMultiplePermissions
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.core.content.ContextCompat
-import com.example.proyecto1.ui.theme.Proyecto1Theme
-import com.example.proyecto1.ui.theme.screens.NetworkMonitor
-
-class NetworkMonitorActivity : ComponentActivity() {
-
-    private lateinit var wifiManager: WifiManager
-    private lateinit var connectivityManager: ConnectivityManager
-    private lateinit var networkMonitor: NetworkMonitor
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        wifiManager = getSystemService(WIFI_SERVICE) as WifiManager
-        connectivityManager = getSystemService(CONNECTIVITY_SERVICE) as ConnectivityManager
-
-        // Creamos la instancia de NetworkMonitor pasando los servicios y la actividad actual
-        networkMonitor = NetworkMonitor(wifiManager, connectivityManager, this)
-
-        setContent {
-            Proyecto1Theme() {
-                Surface(modifier = Modifier.fillMaxSize(), color = MaterialTheme.colorScheme.background) {
-                    networkMonitor.NetworkMonitorScreen() // Se invoca la pantalla de monitoreo
-                }
-            }
-        }
-
-        requestPermissionsIfNeeded() // Solicita permisos al inicio
-    }
-
-    fun requestPermissionsIfNeeded() {
-        val permissions = listOf(
-            Manifest.permission.ACCESS_FINE_LOCATION,
-            Manifest.permission.ACCESS_COARSE_LOCATION
-        ).filter {
-            ContextCompat.checkSelfPermission(this, it) != PackageManager.PERMISSION_GRANTED
-        }
-
-        if (permissions.isNotEmpty()) {
-            requestPermissionsLauncher.launch(permissions.toTypedArray())
-        }
-    }
-
-    val requestPermissionsLauncher =
-        registerForActivityResult(RequestMultiplePermissions()) { permissions ->
-            if (permissions[Manifest.permission.ACCESS_FINE_LOCATION] == true ||
-                permissions[Manifest.permission.ACCESS_COARSE_LOCATION] == true
-            ) {
-                Toast.makeText(this, "Permisos concedidos", Toast.LENGTH_SHORT).show()
-            } else {
-                Toast.makeText(this, "Permisos no concedidos", Toast.LENGTH_SHORT).show()
-            }
-        }
-}
